requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
cryptography>=42.0.0
pytest>=7.4.4
pytest-asyncio>=0.23.2